Implemented Date Parsing and Time Addition
Added the ability to parse a date string according to a specified format.

Added add[Unit] (e.g. addMinutes) methods to easily add units of time to the current date/time. These update the current object but also return the current object to made stringing them together possible.

Changed some reference information in the object to be static for visibility and memory purposes.