Fixes to formatting upper case AM/PM and parsing months/weekdays
Fixed symbol "P" being missing from the list of format characters, resulting in the parser being unable to parse from them or format to them.

Fixed an incorrectly specified array index limit that was preventing a string being parsed when it ended with a word-format month name or day of the week. Additionally, merged the parsing for these two into a single utility function for neatness.