Fixed Undefined Query String Values
Fixed a bug where a value in a query string would parsed as an empty string rather than a null. This was intended so a missing value would be treated as an empty string only if the "=" character was included. null therefore represents where the request explicitly did not intend to supply a value.