Added iteraction count to fInterval and fixed a bug.
Added the option to the "start" method to allow a number of iterations to be specified for the interval. Once the callback has been called that many times, the interval will stop.

Additionally, fixed a bug that could allow the interval to be started multiple times concurrently under most conditions with repeated start() calls. This should no longer be possible.