Added function overloading to javascript.
Added a function "overload", which will accept 'arguments' variable from the calling function, as well as a series of function signatures and functions.

The overload function will check each signature against the parameters in turn and, if a match is found, the associated function is executed. A default function can also be specified at the end, which will be called if all other signatures are invalid.